Remove sighash fork value (forkid)
Addresses #432.

"ForkID" is used to describe the signature hash type which is used after
UAHF, so the majority of references in the code to ForkID will remain -
it's necessary to be able to distinguish between sighashes with and
without a fork ID. The value of the fork ID itself (referred in the code
as a "fork value") is an arbitrary `uint32_t` which has always been zero.
This MR removes the potential to set this to anything other than zero.

Changes
-------

* Remove `SigHashType`'s fork value functions:
  * `withForkValue()`
  * `getForkValue()`
* Remove subsections of test cases concerned with altering and checking fork value:
  * `sighash_tests/sighash_test`
  * `sighashtype_tests/sighash_construction_test`

Out of scope
------------

There are over 200 lines of code remaining which refer to "fork ID", in
one form or another in the source. These relate to the sighash type (all
post hard fork) and are all retained. To avoid reference to an unused
fork id, it may be a good idea to rename these flags in future.

Similarly, the flags `SCRIPT_ENABLE_SIGHASH_FORKID` and
`bitcoinconsensus_SCRIPT_ENABLE_SIGHASH_FORKID` remain as they are used
by logic which decides whether a fork ID should be present or not.
Again, renaming these in future to something related to the HF could be
a good idea.

Test plan
---------

- `ninja check-all`