Replace boost::thread with std::thread
The main complexity here is due to `boost::thread` being interruptible
while `std::thread` is not. In most cases where `boost::interruption_point`
existed, a check against whether or not shutdown is in process already exists,
or has now been added to handle the only relevant interrupt in these places. 
There is some complexity in passing references required to gather the shutdown
status in the case of RPC jobs, however. To be as consistent with Core as 
possible, I have ported across their solution, which involves passing a 
reference to `NodeContext` to each RPC job. From there, the jobs can now fetch
an RPC interruption point callback function that is now attached to 
`NodeContext`.