trivial: Avoid use of inline const non-POD global variables
There is only 1 instance of such a construct in the codebase, in
`amount.h`, added in !1406. It has been reverted back to using `extern
const` and defining it in `amount.cpp`

Rationale: `inline const` of non-trivial objects such as `std::string`
causes the compiler to emit "guard variables" and bloat the emitted
asm code slightly, as well as waste cycles on app startup.

- See:  https://godbolt.org/z/a5jK6fEsK vs https://godbolt.org/z/5nxnh5Kbn