Modify the generic code in `hash.h` to support alternate hashers
This commit modifies the CHashWriter and CHashVerifier classes to be more generic
and to support hashing and verifying to *any* hasher (by default they use the status
quo sha256d).

The motivation for this is that specs such as the utxo-fastsync spec use sha256 *single*
hashing in some places and it was awkward and slow to hash in this way to a vector,
when we could make use of the generic hashers in `hash.h`

Also in this commit -- modified some of the classes in the crypto/ directory
such as class `CSHA256` (the sha256 single-hasher primitive) to support this
codebase's more modern `Span` API (so that these hashers can be dropped-in to
e.g. the `GenericHashWriter` class as a template argument).