Update to DBWrapper API to support taking a leveldb "snapshotted" iterator
This is a part of on-going code changes to support the utxo-fastsync spec.

A snapshotted iterator sees a consistent, "frozen" view of the db in some
past state (at the time the snapshot was taken).  We will need such
a view to be able to dump the txdb for utxo-fastsync data file generation
without the need to hold locks while doing so. The node can continue
to write to the DB in another thread and our custom itertor view of the DB
will remain consistent, until the snapshot is released.

leveldb supports this facility in a thread-safe and performant manner.
When a snapshot's pointer is released, leveldb cleans up any filesystem or
in-memory structures that support the old view.

This commit adds code to optionally create a CDBIterator (and a higher-level
CoinsViewCursor) which is snapshotted.  Existing code remains unchanged
save for this new API extension to our db wrapper classes to optionally
get an iterator to a db snapshot (frozen, read-only view of the db at the
time the iterator was created).

See leveldb docs on snapshots:
https://github.com/google/leveldb/blob/aa5479bbf47e9df86e0afbb89e6246085f22cdd4/include/leveldb/db.h#L102

This commit also adds basic tests that this facility works as expected.