Add ReadRawBlockFromDisk to read raw block data without deserialization
Co-authored-by: Calin Culianu <calin.culianu@gmail.com>

This is a performance improvement as discussed in #424, to read raw block data
directly from disk to the network message, saving on deserialization and
reserialization.

It is implemented through a new `ReadRawBlockFromDisk` method which is tested to
produce the same output as network-serialized block read from `ReadBlockFromDisk`.

A new configuration setting, `-checkblockreads`, has been added to enable extra
sanity checks on the block data read from disk that are to be served up for the
p2p network. This option is intended for testing and may negatively impact
performance.

Test Plan
---------

- `ninja check-all` should pass without errors

Closes #424