[test] Add test harness for UI tests
This MR contains add test framework enhancements for the new UI testing.
It adds a UI test plan GUI and updates the testing documentation with an
example UI test in Python.

Note about UI tests:
The pass/fail status of UI tests cannot be accurately reflected by the
test runner's conclusion. It must be determined by the tester's evaluation
of the test steps in the GUI wallet.

On some platforms, residual stderr output may be generated which causes
the test framework  to mark the test as 'failed' (because tests should
not generate unexpected stderr outputs). This can safely be ignored as
long as the test itself runs "according to plan".

Test plan:
- Run `test/functional/test_runner.py ui-example-test.py`
- Run `ninja check-all`

Closes #438