Add DSP handling in UI wallet
This MR introduces passing DSP signals from core layer to the UI layer.
In the UI layer we add the DSP detection related to the currently opened
wallet. Upon proof arrival, an application notification will be shown,
additional information will be shown in the transaction list and transaction
info window.
To lay out the foundation for UI testing, we also introduce new `extra_args`
flag "-ui", which spawns a UI enabled node, which can be "scripted" with
RPC calls. UI tests shall be denoted with "ui_" or "ui-" prefix and will
be excluded if test runner is tasked to run all tests, because UI tests
are meant to be run separately.

Test plan:
- Run `ninja` to build project
- Run `./test/functional/test_runner.py ui-bchn-rpc-dsproof.py`.
  bitcoin-qt will be launched.
- Switch to transaction tab promptly.
- Wait 10 seconds
- An application notification with "Double Spend Proof" caption will appear.
- Transaction involved in a double spend attempt will become marked as red,
  status becomes exclamation mark.
- Double click on this transaction to view additional info related to double
  spend detected.
- Wait 10 more seconds, transaction will become mined, double spend info
  still available.

Closes #437