Benchmarks: Independent runs for each eval
For full motivation, see #412: Benchmark evaluations are not independent.

This resolves the issue described above, ensuring that each eval is independent.

## Changes

- Evaluations run independently, with all setup executed for each evaluation from scratch.
  Iterations continue to be visible to the benchmarks.
- Benchmark dispatch logic simplified - evaluations and iterations are no longer merged together.
- A `StartBenchmark()` function is added to start the benchmark clock at the
  beginning of the benchmark loop, after individual benchmark setup.
- A `BENCHMARK_LOOP` macro replaces the existing `while (state.KeepRunning())` loops.
    - This allows for additional loop setup to be added at the start, and
      maintained centrally in bench.h.  Future changes can be made without
      further modification to individual benchmarks.
    - This currently calls `StartBenchmark()`.
- Evaluations count member variable is removed from the benchmark state;
  number of evaluations is only visible to the benchmark runner.
- Benchmark state variables intended only for use with benchmark internal
  logic are made private and no longer exposed to the benchmarks, except
  number of iterations.
    - Logic in individual benchmarks that attempted to predict number of executions
      based on evaluations and iterations updated to only see iterations.
- General refactoring and tidying of bench.h and bench.cpp and benchmark state.

## Further work

This change highlights some places in benchmarks where repeat runs are not
truly independent of each other as they should be - for instance
`CCheckQueue_RealBlock_32MB_WithCacheStore`, which fails to properly clear
the sigcache during startup, so subsequent evaluations run much faster than
the first.  This will be addressed in a future MR.

There is some significant inconsistency in the time taken between individual
benchmarks, based mostly on default numbers of iterations.
This is addressed in !1480.

## Test plan

```
ninja bench_bitcoin
src/bench/bench_bitcoin
```

The effect of this change is best observed in combination with MR !1498,
which resolves an issue with persistent caching in the
CCheckQueue_RealBlock_32MB_WithCacheStore benchmark.  If this branch is
merged into !1498 (or vice versa), try:

```
ninja bench_bitcoin
for evals in {1..10}; do src/bench/bench_bitcoin -filter='CCheckQueue_RealBlock_32MB_WithCacheStore' -evals=$evals | grep -v '^#'; done
```

Observe on master `min` reduces drastically as `evals` increase, as setup
cost is counted in the first evaluation only.  The same occurs both on this
branch without !1498, and on !1498 without this branch.  Verify there is
no equivalent effect when this MR and !1498 are both applied.

Example results can be found in the MR 1499 description on GitLab.

Note that the changes in !1498 are needed to demonstrate this effect, as the
signature cache is not correctly reset without them (although the current
benchmark expects it to be).