Define `upgrade10activationtime` for `-chipnet` (was undefined!)
Summary
---

After merge of !1600 (and !1601 & !1604 implicitly), it turns out that
`-chipnet` never got a properly-defined "upgrade 10 activation time". It
turned out to be defined as 0 (which means the `-expire` option won't
take effect luckily), so things work.  However, we really should be
defining an expiry time for this network and not rely on it
"accidentally" ending up a 0 (this is UB technically).

This MR does just that.  After some discussion here:
https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/merge_requests/1601#note_1183736040

.. it was decided to set the upgrade10 time for chipnet to Nov. 15,
2023, 12:00:00 UTC.

Note that for now this option just affects when the node expires via the
`-expire=1` (default: 1) mechanism.

Help was also updated to reflect the differing time(s) for chipnet versus
other networks for both upgrade9 and upgrade10 activations.

Test Plan
---

- `ninja all check-all`
- See help and verify it looks ok: `./src/bitcoind -hh | less` and
  scroll down to the `upgrade9activationtime` option and its sister
  option, `upgrade10activationtime` and review that the times displayed
  are correct (both should be Nov 15 2022 and Nov 15 2023 both at 12:00:00
  UTC respectively)