Update checkpoints for mainnet, testnet3, testnet4, and chipnet
Summary
---

In preparation for release, we are updating the compiled-in checkpoints
for mainnet and for 3 of the 5 testnets.  (ScaleNet and RegTestNet never
get a checkpoint update for differing reasons).

Test Plan
---

Use a blockchain explorer or a node on each of the 4 networks that got
new checkpoints to determine that the added checkpoint blocks are indeed
on the current active chain at the height specified.

For reference, a chipnet explorer to use is: https://chipnet.chaingraph.cash/
A suite of main, test, test3/4 explorers is available here: https://bch.loping.net/