Update chain params (assume valid and minimum chain work)
Summary
---

As per the release process, we update the assume valid and minimum chain
work chain params to recent values, using the process described in
`contrib/devtools/chainparams/README.md`

Test Plan
---

For each of the 4 chains updated (mainnet, testnet3, testnet4, chipnet),
verify that the assume valid block is on the active chain and it should
be relatively recent (~2000 blocks from the current tip as of the time
of this MR for testnets, ~10 blocks from the current tip as of the time
of this MR for mainnet).

Also verify that the minimum chain work is close to the current chain
tip's work but does not exceed it (interpret this hash as if it were a
uint256 arithmetic value, big endian byte order).