Update seeds for mainnet and testnet4, plus fix bug in generare-seeds.py
Summary
---

Followed the instructions in contrib/seeds/README.md, however I lack
a seeder dump from testnet3, so this only updates seeds for mainnet
and testnet4. Chipnet was recently updated in !1600 and doesn't really
need an update at this time. It does delete the 1 no-longer-valid
.onion address from testnet3.

Also in this commit: We needed to update the generate-seeds.py script
to:

1. Accept new longer .onion addresses and reject old ones.
2. Use the new #pragma once for generating chainparamsseeds.h

Test Plan
---

- Review
- (Optional) Check some of the new additions actually are nodes that are up
  (note that nodes may be down temporarily especially the .onion ones).