Update chainTxData for main, test3, test4, scalenet, and chipnet
Summary
---

This information is used in some RPCs to display "verification progress"
which is basically just advisory.  Currently verification progress is
always off anyway since it's such a rough guess, but as part of the release
process it's good to update this from time-to-time.

The method used was to check the `UpdateTip` line from the debug.log to
get the latest nTx count for the chain, and then get that block's
declared block header time, and do some calculations from the last time
we sampled this information to derive the guesstimated tx rate.

This was done for each chain.

Test Plan
---

- `ninja all check`
- (optional) review that the numbers are correct for the declared blocks
  and heights on each chain (may be tricky to get unless you run a node
  or use one of the explorers at bch.loping.net).