tests: Add RPC-based basic tests to send/receive tokens to/from wallet
Summary
---

This tests a few RPCs that offer very basic token support (such as
wallet `listunspent`) and token signing of wallet txns, sending of
tokens to a wallet, and ensuring that the wallet can "see" the token
utxos but by default never selects them for spending.

For now, the user of the RPC wallet must manually construct transactions
using `createrawtransaction` and its `psbt` cousin, if they wish
to actually send tokens from a wallet utxo.  The wallet *will* know
how to sign for such transactions (even if it refuses to construct
token-sending txns itself).

Note that the `getbalance` RPC *does* take token UTXOs into account
even if it can't actually select them for spending.  This was
intentional so that the user knows "something funny is going on" and
hopefully investigates why they have additional hidden balance after
they receive a token.

If the above behavior is objectionable, this can be modified (up for
discussion).

Also in this MR:

- Added a function to the framework to calculate dust limit accurately.
  This is increasingly important since token-containing CTxOuts have
  a variable length now and the dust limit is not always 546 for them --
  sometimes it may be as high as ~800!
- Fixed a bug in deserialization of TokenOutputData in the test
  framework itself which was not deserializing token data correctly.

Test Plan
---

- `ninja all check-all`