Fix bug: RPC & REST API sometimes reads from ::ChainActive() without locks
Summary
---

This closes issue #465.  Some parts of the codebase were reading from
::ChainActive() without `cs_main` held. This was happening in particular
in the REST and RPC APIs (calls such as `getblock`).

Core and ABC fixed these bugs in some of their many refactors, but we have
diverged from them and it is easier just to plug this bug now than to
try and synch up with them and fix this bug as a side-effect.

The approach used to fix this was approach (2) discussed in #465. One
could have also done approach (1), that is, annotate `::ChainActive()`
with an `EXCLUSIVE_LOCKS_REQUIRED(cs_main)` macro but the diff for that
would be rather large and mostly filled with boilerplate to get tests
working.

It is hoped by the author of this MR that by keeping this diff small it
gets merged more quickly into this codebase, since it really is a bug.

Test Plan
---

- `ninja check-all`