Update checkpoints for v26.0.0 (main, test, test4, and chip)
Summary
---

Update the checkpoints to a recent block for the 4 major testnets (minus
scalenet).  Main was updated to ~12 blocks ago as of the time of this MR
and the testnets to ~1000 blocks ago as of the time of this MR.

Test Plan
---

Verify that the blockhashes for each of the 4 testnets provided are
indeed on the main active chain for that network, at the specified
height.