Update chainparamsconstants.h for version v26.0.0
Summary
---

As part of the release process, we update the `chainparamsconstants.h`
file using the instructions in `contrib/devtools/chainparams/README.md`.

"Assumvalid" affects whether signature verification is not done before
the block hash in question, and minimum chain work affects whether we
elect to reorg or not should there be a chain split after the release
(unlikely).

For mainnet assumevalid was taken 10 blocks from the tip (as of the time
of this MR), and for testnets it was 2000 blocks from the tip.

Chainwork was taken from the tip as of the time of this MR.

We don't update these parameters for scalenet since it is expected to be
reorged down to block 9999 someday.

Note: The comment that was at the top of `chainparamsconstants.h` was
removed automatically by the script that re-generated this file.

Test Plan
---

1. Verify that the block hashes in question are indeed on the active
   chain for the networks in question.
2. Verify that the "minimum chain work" specified is <= the current
   chain tip's work for the active chain tip on the networks in question. It
   should be close to the value of the tip but not exceed it.