Prevent theoretical UB for CConnMan and PeerLogicValidation mis-use in tests
Summary
---

This commit fixes UB I noticed when backporting for MR !1697.
It is a commit on top of !1697 and depends on it.

This fixes issue #474. For a description of why this commit is
needed, please see #474.

The strategy employed here is to ensure the lambda passed to the
scheduler in the `PeerLogicValidation` and `CConnMan` code un-registers
itself from the scheduler should the lifetime of its `this` instance end.
We accomplish this by using a `std::shared_ptr<std::atomic_bool>` to
tell the lambda that the class instance it operates on has died.

If it has died, the lambda returns false, thus getting itself deleted
from the scheduler for eternity and never executing again. (At which
point the atomic_bool shared_ptr is also deleted).

Test Plan
---

- `ninja all check-all`