Fix potential UB in wallet test fixture
Summary
---

Attempt to fix further UB after investigating issue #472.  The purpose
of this commit is to safeguard the code against future UB, by making it
safer.

The basic gist of the pathology is:

1. Wallet test fixture tests may have "validation interface" work
   scheduled to run in the wallet context (they are enqueued for the
   scheduler's single threaded process queue, as a result of validation
   interface signals like "block connected").
2. The wallet is deleted by `WalletTestingSetup` d'tor.
3. This may sometimes happen at the exact same time as a scheduler
   callback is delivered to the (now deleted) wallet.
4. UB ensues!  Note this only affects the tests. The main app has the
   fortunate property of stopping the scheduler first before doing any
   destruction of objects, wallet unloads, etc.

Currently no tests should be enqueuing validation interface signals
however, but the corner case should still be handled elegantly.

The fix was to provide a service method to the `TestingSetup` superclass
called `StopScheduler()` to stop the scheduler early, before we delete the
temp wallet used by `WalletTestingSetup`.

Test Plan
---

- `ninja all check`