[backport] Rate-limit unsolicited p2p ADDR messages
Co-authored-by: Jon Atack <jon@atack.com>
Co-authored-by: Calin Culianu <calin.culianu@gmail.com>

Summary
---

This is a backport of core#22387 and core#22604 smooshed into 1 commit.
It was adapted slightly to our codebase since out p2p network data
structures are different than core's (we didn't do all of their noisy
refactors), so some details of the code changes differ from Core PRs
slightly, but the functionality is identical.

The motivation for this backport is that it is a decent performance
optimization to rate-limit unsolicited ADDR messages from peers in order
to minimize on wasted CPU cycles from hypothetical badly-behaved peers.

Original Core PR description:

    The rate at which IP addresses are rumoured (through ADDR and ADDRV2
    messages) on the network seems to vary from 0 for some non-participating
    nodes, to 0.005-0.025 addr/s for recent Bitcoin Core nodes. However, the
    current codebase will happily accept and process an effectively unbounded
    rate from attackers. There are measures to limit the influence attackers
    can have on the addrman database (bucket restrictions based on source IPs),
    but still - there is no need to permit them to feed us addresses at a rate
    that's orders of magnitude larger than what is common on the network today,
    especially as it will cause us to spam our peers too.

    This PR implements a token bucket based rate limiter, allowing an average of
    0.1 addr/s per connection, with bursts up to 1000 addresses at once.
    Whitelisted peers as well as responses to GETADDR requests are exempt from
    the limit. New connections start with 1 token, so as to not interfere with the
    common practice of peers' self-announcement.

See the original core PRs for more details:

- https://github.com/bitcoin/bitcoin/pull/22387
- https://github.com/bitcoin/bitcoin/pull/22604

Test Plan
---

- `ninja check-all`