Performance fix: Modify logic slightly for the msghand thread sleeptime calc.
Summary
---

While reviewing the network code, I noticed that a particular corner
case can arise where nodes that are not fully connected or that are
flagged for disconnection are taken into consideration for the "minimal
sleep time" for the message handler thread.  These nodes will never ever
send invs (if you read PeerLogicValidation::SendMessages() you will see
why).  As such, a minor performance optimization would be to not take
these nodes into account for calculating the minimal sleep time for this
thread's cond. wait, since the cond. wait time is a function of when we
anticipate next needing to send invs.. and these nodes can never send
invs, so they should not be in the calculation.

Test Plan
---

- `ninja all check-all`