Remove repeated calls to the slow gArgs object in perf. critical network code
Summary
---

I did some profiling of the code of `PeerLogicValidation::SendMessages`
and this function ends up being called at least 10 times per second, and
sometimes as many as 100 or 1000 times per second during periods of very
high network activity.  Usually this function returns very quickly, but
surprisingly, my profiler showed me that the cost of the call is
dominated by the call to `gArgs.GetBoolArg("-feefilter", ...)` which parses
a string behind the scenes to return a bool. The "-feefilter" argument
never changes throughout the run of the program, not even in tests, and
thus can just be a member variable for `PeerLogicValidation` set at
construction time by `init.cpp` (coming from this CLI arg).

Thus, we do not need to parse the argument string each time that
`PeerLogicValidation::SendMessages()` is called as this is wasteful, and
we just cache it to a bool member variable.

In my profiling here on macOS with clang and -O3 -g, the profiler showed
a 30% cost reduction to the cost of this function's execution time (this
was the single most expensive thing in the function).

Test Plan
---

- `ninja all check-all`