Remove redundant calls to `GetTime*()` in msghand thread
Summary
---

During profiling, it was revealed that `PeerLogicValidation::SendMessages` is
sometimes called extremely frequently.  One of the top performance
bottlenecks in this function, believe it or not, is the calls to
`GetTimeMicros()` and `GetTime()`.

As such, we can reduce these calls by taking the "current time" at the top of
the function once, and then just using the time saved to that variable. We must
take the current time as both a mockable time (`GetTime()`) and
non-mockable (`GetTimeMicros()`).  But we take it once and we shave off
some CPU cycles by doing this.

I checked and Core already refactored this and is doing similar. They
are ok with taking the timestamp once up top and then they use that.

FYI - The reason why `GetTime*` and friends are slow-ish is
that they ultimately call into boost which calls into C library
`gmtime()` function which is kind of heavy because it fills in a big C
struct each time.

Test Plan
---

- `ninja all check-all`