Refactor and docs for ProcessGetData
This is a small refactor for `ProcessGetData`.

The while-loop now deletes processed items as-it-goes, instead of one single
erase at the end. This is more idiomatic to this codebase, and easier to 
reason about for most people.

Also, a condition to break out of the while-loop is added to the 
while-condition where it belongs (instead of the body of the while-loop with 
a `break`). The semantics of the loop are changed in a very small way because
`interruptMsgProc` will be checked in a different order. This is immaterial 
to the general execution of `msghand`'s flow.

Lastly, the flow of the function and the "whys" of it are explained in a 
comment at the top.

---

TEST PLAN

```
ninja check-all
```