Fix issue with peers not receiving DSProofs if they ever use the mempool p2p message
Summary
---

This fixes issue #481. The pathology is due to a coding/logic error that
was originally introduced with the dsproof subsystem. The problem is
that if a peer ever requests the mempool, they will never receive a
dsproof due to the way the code is written.  See issue #481.

The functional test `bchn-feature-doublespend-proof.py` was also updated
to catch this regression from ever occurring again.

Test Plan
---

- `ninja all check-all`