[qt] Improve importwallet progress display
This changes the progress bar display during 'importwallet' in the
GUI as follows:

First, a separate 'Scanning...' progress bar is displayed while the
dump file is being parsed to count the number of keys and scripts.

A second progress bar is created for the actual key/script
insertions.

A new debug log output is created for the case of imported scripts
(previously only new keys and errors were logged). This is to
facilitate checking of the correct function.

## Test Plan

* 'ninja all check-all'
* `test/functional/test_runner.py ui-importwallet-progress`

Follow the prompted steps of the GUI test framework.

Please note that the standard test harness will still log GUI tests
as 'Failed' when they finish. What matters to this test plan is
whether the steps in the GUI test prompts produce the expected results.