[backport] Create blockstorage module
This is a refactor that creates a new blockstorage module based on Core's
PR21575 and PR21727:

- https://github.com/bitcoin/bitcoin/pull/21575
- https://github.com/bitcoin/bitcoin/pull/21727

It moves functions that deal with disk IO for block data into this new module.
This helps cut down the noise in init.cpp and validation.cpp and should
simplify access to disk functions for testing/benchmarking.
No new functionality is introduced.

# Test plan

- `ninja check-all`
- While checking the diff, see moved lines more easily with, eg:
  `git diff master --color-moved=dimmed-zebra --color-moved-ws=ignore-all-space`