Modify method `HTTPReply::WriteReply` to use `Span<const uint8_t>`
Summary
---

`HTTPReply::WriteReply` accepts a `const std::string &`. In many places
in the codebase, especially for heavier objects, this required building
a temporary string from a different buffer, which led to extra
allocations and copying.

This commit fixes the situation by having this method accept a zero-copy
`Span<const uint8_t>`, so that it will work much faster with binary
data. Additionally, a `std::string_view` overload has been provided
which also is zero-copy and will work conveniently for places wishing to
write a string to the HTTP reply.

Closes issue #494.

Test Plan
---

- `ninja all check-all`