Optimize two block reads with raw version
Two instances of `ReadBlockFromDisk` are replaced with `ReadRawBlockFromDisk`. 
The former would unserialize the block to `CBlock` representation and then
reserialize it back into the network, while the latter will read the raw block
bytes which are already the serialized representation.

Assumption: Network serialized representation is the same as Disk serialized
representation. Caveat: the codebase contemplates the eventuality that the
two serialized representations may not match, see: 
https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/blob/master/src/serialize.h#L166

The first instance is in `src/zmq/zmqpublishnotifier.cpp`, which was a simple
unserialize and reserialize. The second instance in `src/rest.cpp` is a little
more complex, as there are three cases (binary, hex and json) and only the third
requires an unserialized representation: the block is read raw and used raw for
the first two, and unserialized as needed for the third.

---

TEST PLAN

```
cmake -GNinja .. -DBUILD_BITCOIN_ZMQ=ON
ninja all
ninja check-all
pip install zmq
test/functional/test_runner.py interface_zmq.py
test/functional/test_runner.py interface_rest.py
```

The last two functional tests in particular are already implemented to catch
errors in representation.