trivial fix: Ensure HeapOptional move-constructor actually moves
Summary
---

While looking over the code, I notice there is a small error in
`util/heapoptional.h`.  The "move" constructor from a type `T &&` should
be calling `std::move(t)` again when invoking operator=, but it was not.

This causes `operator=(const T &)` to be called instead of the
potentially faster `operator=(T &&)`.

This commit fixes the above to call the correct `operator=`.

No real behavioral changes are introduced by this commit, it is just a
nit-commit.

Test Plan
---

- `ninja all check-all`