Remove "zero after free" allocator from codebase (used by CDataStream)
Summary
---

This MR removes the `zero_after_free` allocator which is used by the
`CDataStream` class's own internal buffer. The `CDataStream` class is
used everywhere for serialization/deserialization, especially in the
network code.

This allocator was used to zero memory after freeing it.

With this MR, we remove it since it does not have a strong reason to
exist, and instead it just eats CPU cycles.

The existing `secure_allocator` (which is used in parts of the wallet or
in code that manages keys) is left in-place and it has a reason to exist.
The `zero_after_free` allocator really does not.

Zeroing memory pages after freeing them is not particularly useful if
processing public data (as the network code does).

The rationale for originally having CDataStream work this way has an
interesting history, which you can read about in this core PR:

https://github.com/bitcoin/bitcoin/pull/740#issuecomment-3356239

The above-linked comment makes an argument that zeroing memory helps
reduce the potential damage of buffer overflow bugs. The argument
doesn't hold much water, in my opinion.  Since this is C++ and we use
vectors everywhere, this type of bug has a near 0 proability of existing.

However, assuming such a bug were to exist, `zero_after_free` doesn't
actually prevent these bugs from doing infinite damage anyway, or from
existing in the first place. If such a bug were to exist remote code
execution is possible anyway so it's not like zeroing pages buys you much
in that regard. In other words, if you assume that remote attacker code
execution is possible, zeroing pages doesn't save you.

In short, zero_after_free for general purpose code that is working with
non-key data is just paying a non-zero performance cost for zero actual
benefit.

Test Plan
---

- `ninja all check-all`

Unfortunately no benchmarks exist for the `CDataStream` class, really.
One would expect a performance speed-up from removing the allocator,
however.