Fix for failing unit test
Summary
---

This unit test was testing bip69 sorting of txns in the wallet, and it
was creating out-of-consensus CashToken txns.  This test therefore
stopped working once the upgrade9 activation time had elapsed today.

Since the point of the test is to just test that txs sort correctly, and
nothing more, we simply "fixed" the test by ensuring upgrade9 is NOT
activated for this test.

Rationale: It would be too complicated to fix this test to actually create
legit cashtoken txns, when all it is doing is testing bip69 sorting anyway.

FWIW there is another test in this file that is actually testing the
upgrade and it has CashTokens and works with upgrade9.

Test Plan
---

- `ninja all check-all`