Add support for generic serialization of std::optional<T>
Summary
---

This commit adds generic serialization/unserialization support for
any std::optional<T>, if T is already serializable/unserializable.

The raw serialization format for optionals is to simply prepend a single
byte to the stream to indicate if the optional value is present (0 not
present, 1 present), and then proceed to ser/unser the T to/from the
stream if the optional is present (and do nothing more if it is not).

Added unit test for std::optional serialization as well.

Test Plan
---

- ninja all check