Add ReadBlockSizeFromDisk() function to blockstorage.cpp
Summary
---

We add a function to `blockstorage.cpp` & `.h` called
`ReadBlockSizeFromDisk()`.  It returns just the block size of any
particular block that we have data for.  It is a fast function to call
and we may need it for up-coming EBAA work.

The existing function `ReadRawBlockDromDisk()` was refactored, and it now
calls `ReadBlockSizeFromDiskCommon()` which is a private function that
does the actual block size reading, and it is called by both the
existing `ReadRawBlockFromDisk()` and the new `ReadBlockSizeFromDisk()`
function.

Unit tests were updated to test the new `ReadBlockSizeFromDisk()`
function.

Test Plan
---

- `ninja all check-all`