Added new CLI arg `-percentblockmaxsize` as an alternative to `-blockmaxsize`
Summary
---

A new CLI arg `-percentblockmaxsize=<percent>` was added, where
`<percent>` is a percentage value in the range [0.0, 100.0].

This new CLI arg is potentially useful in the future when we will have a
dynamic block size limit, because some miners may wish to cap their
mined size as a fraction of the dynamic block size limit, rather than as
a static byte count.

The argument is expressed as a floating point percentage value, and the
default for mainnet is 25.0 which corresponds to 8000000 (status quo
default mined block size).

Both `-blockmaxsize` and `-percentblockmaxsize` cannot be specified at
the same time.

If the percentage argument was specified, then the percentage is remembered
and a future hypothetical dynamic block size algorithm will use this
percentage for mining, thus the mined block cap will grow as the consensus
block limit grows.

If a byte count is specified, then a byte count is remembered and any
blocksize AA won't scale the mined block size.

If nothing is specified, then the node defaults to percentage mode
(rather than byte count mode) and it *will* scale the generated block
size with a scaling of the max consensus block size.

Note
---

Note that a slight semantic change was added with this facility. The
node now "defaults" internally to using a chain-specific *percentage*
for the default max generated block size (mainnet: 25.0, testnet: 25.0,
testnet4: 100.0, scalenet: 3.125, chipnet: 100.0, regtest: 25.0).

This, if the user doesn't set `-blockmaxsize` but sets
`-excessiveblocksize=xxx`, then e.g. for mainnet the node will default
to using 25% of `xxx`. Before this change, it would default to using 8 MB
regardless of what the user specified for `-excessiveblocksize` (if they
didn't also specify `-blockmaxsize`).

The rationale for this is we want the node to behave intelligently "by
default"... and this new semantic change does future-proof the code for
when we add a dynamic blocksize adjustment algorithm to the codebase.

Test Plan
---

- `ninja all check-all`