Add the ability to see the configured `-blockmaxsize` from the RPC
Summary
---

There is no way to query the node to discover its configured
`-blockmaxsize` value via the RPC. As such, this commit adds this
ability. A new key was added to the `getmininginfo` RPC called
`miningblocksizelimit`, which basically is the same as the chain's
default mining limit and/or the `-blockmaxsize`, if configured by the
user.

The release notes were updated as were tests to test for this.

Test Plan
---

- `ninja all check-all`
- Try starting the node with various values for `-blockmaxsize`
  (including unspecified), and verify that the `getmininginfo` RPC key
  `miningblocksizelimit` matches what is expected.