Remove `const_cast<Config &>` from various places in code.
Summary
---

This is a bit of a code smell. Also technically `const_cast` away of
`const` and then writing to the object *may* lead to UB, according to
the standard. See:  https://stackoverflow.com/questions/4042626/const-cast-and-ub

In practice it won't cause problems but it is still not ideal to see.
The code in question does setup the `Config` object anyway (from CLI
args, etc), so it should really just be working with a non-`const`
`Config` object, not a `const` one.

Test Plan
---

- `ninja all check-all`