test framework: Pass down the rpc_timeout to wait_until on node restart
Summary
---

In the python test framework, when restarting the node (by calling e.g.
`self.restart_node()`), `wait_until()` is used to probe the node that it
is fully restarted by checking that the `mempool.dat` file has finished
loading.

This `wait_until()` call is using a timeout of 60 seconds always, even
if the test in question specified a longer `self.rpc_timeout`.

For the upcoming ABLA work, on the sanitizer-undefined CI jobs, the abla
test was failing because it took a very long time for the node to
re-load its mempool (something like 100-200 seconds).  Thus, the static
60 seconds used by the `wait_until()` call used when restarting the node
is inadequate and inflexible.

This commit addresses the issue by passing down `TestNode.rpc_timeout`
to `wait_until()` on node restart.

Existing tests that use the default timeout of 60 seconds are
unaffected.

New tests that specify a `self.rpc_timeout` will benefit from this
change.

Test Plan
---

- `ninja all check-all`