Rename the "excessive block size" concept in the C++ codebase
With the upcoming ABLA changes, the "excessive block size" concept in
the codebase is perhaps confusingly named. After ABLA activates, it no
longer has the static meaning it once had, and as such, it should be
renamed to avoid confusion.

This MR does just that, renames the concept to "consensus block size"
in all places.

Note that the -excessiveblocksize= config arg remains untouched and the
getexcessiveblock RPC call also remains untouched in order to not break
external compatibility.

This is a renaming MR and does not modify functionality in any way.

Test Plan
---

`ninja all check-all`