Add `GetNextBlockSizeLimit()`, refactor `BlockAssembler` to use it
We add a function that will be needed for ABLA, GetNextBlockSizeLimit()
which given a pindexPrev, can tell you the following block's size limit.
Currently it just returns the static configured size
(e.g. via -excessiveblocksize=xxx).

The BlockAssembler class that lives in miner.cpp was refactored to use
this new function, as were various bits of the codebase.

In addition the BlockAssembler was refactored:

- To no longer use a static local for IncrementExtraNonce() and no longer
  require cs_main for this function as a result.
- CreateNewBlock() is now more correct in that it returns what tip it is
  mining against, thus avoiding a potential race condition if called
  without cs_main held.
- Some misc. code cleanup/nits/fixups.

Auxiliary code and tests were updated to use the changed BlockAssembler API.

This is a pure refactor and is not intended to modify external functionality
in any way.

It merely paves the way for upcoming ABLA-related changes.

Test Plan
---

- `ninja all check-all`