Compiler warning fix: Remove unused variable in policy.cpp
After commit 9331b629aa94300ec12c4d26827c6a6d32c08141 , the variable
`nDataOut` was no longer being used (read from). It was only being
written-to because previous to the above-mentioned commit it was used to
enforce relay policy.  But after the above commit, it stopped being
read-from.

On my latest clang it was producing a compile-time warning. This commit
fixes that.