Compiler warning fix: Remove unused variable in policy.cpp
After commit 051e716caeb2ad8085958e632003f0b6b2196c32 , the variable
`nDataOut` was no longer being used (read from). It was only being
written-to because previous to the above-mentioned commit it was used to
enforce relay policy.  But after the above commit, it stopped being
read-from.

On my latest clang it was producing a compile-time warning. This commit
fixes that.