seeder: Gracefully exit and always persist db immediately on exit
The app can now be exited in the usual unix ways (SIGTEM, SIGINT, etc).

The app now fully cleans itself up on exit, and dumps stats to disk
immediately on exit, as opposed to once per hour only (the once per
hour thing still occurs in the background as well, however).

This issue was also something Sipa wanted to fix, see here:
https://github.com/sipa/bitcoin-seeder/issues/76

Note that the extant `ThreadDumper` has an issue when it is saving in
the background: https://github.com/sipa/bitcoin-seeder/issues/108.

A side-effect of this commit is that when the app exits and it does the
final save to disk, no nodes will be lost from the DB, partially
resolving the above issue.

Summary of changes:

- All sleeping in the code (including poll/select) now periodically
  polls the newly-introduced "shutdown requested" flag.
- Added asynch-signal-safe handling of app shutdown signals (SIGINT,
  etc)
- On app signal, set the shutdown requested flag.
- ThreadCrawler socket select/poll now periodically wakes up to check
  the "shutdown requested" flag (this is to allow for fast thread exit
  from the crawler threads).

(cherry picked from commit 9e511fef309e3f5bda1d5e968a50af5acb7b777a)