net: Future-proof and prevent UB in CConnMan::SocketSendData for >2GiB messages
Summary
---

This commit fixes a potential foot-gun and future-proofs
`CConnMan::SocketSendData` should we decide to support >2GiB p2p messages.

The changes allow for `send()` to be called with the correct size type and
also we now clamp the amount of data to be sent to not overflow the length
parameter to `send()`.

In cases where the length parameter would have overflowed, the behavior
would be that the data will be sent in two parts and the sending will
resume later from where it left off.

Since currently >2GiB message sizes do not exist, no behavioral change is
introduced by this code. This is purely a type safety/code-quality commit
to future-proof the code for eventual >2GiB message lengths and at least
prevent UB should such a change be introduced to the codebase.

More Background
---

Currently, the code in `CConnMan::SocketSendData()` would fail if it were
to try and send messages >2GiB (larger than a 32-bit signed int).

This isn't currently a problem since the maximum p2p message length is
bounded by a 32-bit integer currently.

However, should we extend the p2p protocol someday to support >2GiB
message sizes (say, in order to be able to send >2GiB blocks via p2p),
the code in `CConnMan::SocketSendData()` would produce UB and fail.

This commit plugs this "hole" in the codebase and prepares the low-level
sending code to at least not break with UB and support >2GiB message sizes.

Even More Background
---

The reason the original code used `int` here was that it was written for
Windows's winsock library, which is limited to `int` data types for
lengths for `send()`.  However, using `int` here is incorrect because on
64-bit systems, we would overflow the `int` if attempting to send an larger
amount of data than `MAX_INT`.

The changes to the code are such that the size type of `send()` is
detected at compile-time and the amount of data to be sent for each
`send()` call is clamped to not overflow the size type that `send()` is
expecting, thus future-proofing the code should we decide to extend the
p2p protocol to support >2GiB message sizes.

Test Plan
---

- `ninja all check-all`

(cherry picked from commit 5f0e05323c63bdd417651809839b490b55be5978)