Bugfix: Fix potential UB in AddToCompactExtraTransactions in net_processing.cpp
Summary
---

This was taken from !1831:

If -blockreconstructionextratxn=-1 (or any negative value, really) would be
specified in the conf file, the node would generate an exception (or
potentially even crash) when processing messages if it received an orphan tx.

This looks like potential UB which could lead to a crash, and has been
addressed.

The bug occurs because GetArgs returns int64_t but this was being
assigned to a size_t (unsigned type). If -1, then this would be 2^63-1.

Then when the node attempts to resize the vExtraTxnForCompact vector,
a failed allocation would lead to an exception being thrown

Also fixed: more correct to take the iterator position modulo the vector
size, rather than rely on `max_extra_txn` which is not guaranteed to
match in some imagined corner case where args are changed at runtime
(during tests).

No behavioral changes to the code -- this is purely a code quality /
bugfix commit.

Test Plan
---

- `ninja all check-all`

(cherry picked from commit 861aeb647323dbb90a2d08d0f81e1cf18357b088)