net: Fix potential for UB in `CNode::ReceiveMsgBytes`
`CNode::ReceiveMsgBytes` was accessing `hdr.pchCommand` directly as if
it were a NUL-terminated C-string, which is not. This can lead to UB if
the entire 12-byte buffer is full (which is the case for `dsproof-beta`
p2p messages, which are 12-bytes without the NUL).

So, for p2p message `dsproof-beta`, which fills the entire 12-byte buffer,
there was no terminating NUL character when treating this string as a
C-string. This was leading to potential for C++ UB by reading past
the end of the buffer when casting this to a `std::string`.

The correct thing to do is call `hdr.GetCommand()` which transforms
`pchCommand` into a `std::string` correctly using `strnlen`.

In practice this minor bug did not lead to any crashes or any major
issues, just that `getpeerinfo` message counts for receive `dsproof-beta`
messages was always wrong .. `dsproof-beta` was being mis-categorized as
`*other*` in the receive message counts!

This commit fixes the UB and also amends the functional test to detect
correct behavior of `getpeerinfo` message counts (to catch this issue
should it ever regress again).

(cherry picked from commit 5ac80029444ad7c065206903064c35f88d056a22)