Trivial: remove #inlcude <cstdbool> in seeder/dns.cpp
This header is deprecated in C++17 and produces warnings if included on
newer compilers.

It is not needed since it's an old C header that just defines what we
already have in C++: `true`, `false` and type `bool`.

On GCC 15.1.1 it warns, so it needs to go.

(cherry picked from commit 327bb0ce0478cc6eb42f04bf62ce3c280afa3a02)